# Conflicts:
#	src/MoveGenerator.h

+++ src/SearchStats.h
@@ -42,7 +42,7 @@ public:
   int currentSearchDepth = 0;
   int currentExtraSearchDepth = 0;
   Move currentRootMove = NOMOVE;
-  int currentRootMoveNumber = 0;
+  unsigned long currentRootMoveNumber = 0;
   int bestMoveChanges = 0;
   MilliSec lastSearchTime = 0;
